requirements.txt
pandas
fastapi>=0.110.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
google-genai>=0.7.0
langchain>=0.2.0
langchain-ollama>=0.1.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
sentence-transformers>=3.0.0
transformers>=4.41.0
umap-learn>=0.5.0
hdbscan>=0.8.0
scikit-learn>=1.4.0
Sastrawi>=1.0.1
